+++ tests/.pytest_cache/v/cache/lastfailed
@@ -1,3 +1,4 @@
 {
+  "TestFlaskAppRequests.py::TestFlaskUserRequests::test_api_documentation": true,
   "TestFlaskAppRequests.py::TestFlaskUserRequests::test_api_requests": true
 }
